backend/requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-documentai
google-cloud-firestore
google-cloud-bigquery
google-cloud-storage
pydantic
google-cloud-aiplatform
